requirements.txt
tensorflow
opencv-python
numpy
matplotlib
scikit-learn
joblib
Pillow
seaborn
scipy